[Refactor] DTO 클래스 명 변경
