+++ src/main/java/com/bookpub/bookpubfront/member/service/MemberService.java
@@ -1,7 +1,7 @@
 package com.bookpub.bookpubfront.member.service;
 
-import com.bookpub.bookpubfront.member.dto.SignupMemberRequestDto;
 import com.bookpub.bookpubfront.member.dto.SignupMemberResponseDto;
+import com.bookpub.bookpubfront.member.dto.SignupMemberRequestDto;
 
 /**
  * 멤버 도메인의 서비스.
@@ -14,10 +14,10 @@ public interface MemberService {
     /**
      * member 패스워드 암호화 후 shop 서버에 저장 요청하는 메소드.
      *
-     * @param signupMemberResponseDto 멤버 정보를 담아 요청보내는 DTO
+     * @param signupMemberRequestDto 멤버 정보를 담아 요청보내는 DTO
      * @return 저장된 멤버 정보의 일부를 반환받는 DTO
      */
-    SignupMemberRequestDto signup(SignupMemberResponseDto signupMemberResponseDto);
+    SignupMemberResponseDto signup(SignupMemberRequestDto signupMemberRequestDto);
 
 
 }
